improve cgit clone-prefix
Either let them configure on your own or improve how they are
generated from a gitolite perspective.
The basic idea is that these domainalias pre-/suffixes are for
domain names that are under a special ssl certifcate and hence
we want to match the https url that one. All the others should
go with the standard url.